3コラム一覧途中まで

+++ events/show.php
@@ -20,13 +20,14 @@ $user_name = getLoginUserName();
 
 //var_dump('GET', $_GET);
 $event_id = $_GET['event_id'];
-
+/*
 var_dump(
     "event_id = " . $event_id,
     "get['event_id'] = " . $_GET['event_id'],
     "user_id = " . $user_id,
     'user_name = ' . $user_name
 );
+*/
 
 $database_handler = getDatabaseConnection();
 $sql = $database_handler->prepare("SELECT * FROM events WHERE user_id = :user_id AND id = :event_id ORDER BY updated_at DESC");
@@ -73,16 +74,16 @@ while ($result = $sql->fetch(PDO::FETCH_ASSOC)) {
                 <?php  } ?>
             </tbody>
         </table>
-        <a href="edit.php?event_id=<?php echo $event['id']; ?>" class="btn btn-success btn-lg" role="button" aria-pressed="true">この出来事を元に3コラムを作成</a>
+        <a href="../threecolumns/create.php?event_id=<?php echo $event['id']; ?>" class="btn btn-success btn-lg" role="button" onclick="confirmDelete();return false;" aria-pressed="true">この出来事を元に3コラムを作成</a>
         <a href="edit.php?event_id=<?php echo $event['id']; ?>" class="btn btn-primary btn-lg" role="button" aria-pressed="true">編集</a>
-        <a href="delete.php?event_id=<?php echo $event['id']; ?>" class="btn btn-danger btn-lg" role="button" aria-pressed="true">削除</a>
+        <a href="delete.php?event_id=<?php echo $event['id']; ?>" class="btn btn-danger btn-lg" role="button" onclick="confirmDelete();return false;" aria-pressed="true">削除</a>
     </div>
 
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.6/umd/popper.min.js"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/js/bootstrap.min.js"></script>
     <script defer src="https://use.fontawesome.com/releases/v5.7.2/js/all.js"></script>
-    <script type="text/javascript" src="main.js"></script>
+    <script type="text/javascript" src="../public/js/main.js"></script>
 </body>
 
 </html>
